requirements.txt
python-telegram-bot==22.3
pymongo
mysql-connector-python
python-dotenv
